Default to API key for Watson if no ID is set
Watson used to give out an id and password when you requested an account, and those still work.  But nowadays, they just give out an API key.  We need to support both old and new paths here.  So if no id is specified in the Kubernetes secret (exposed as a WATSON_ID environment variable), then assume the API key path (passing in that API key value as the WATSON_PWD env var).