Made Watson ID optional (default to "apikey")
Watson used to give out an id/pwd for each account.  Need to still support that.  But today, they only kind of account you can create uses an API key instead.  in that case, only the "password" field is required (containing the value of the API key), and we'll just assume the id is the string "apikey" if not otherwise specified.